StaticArrays and type stability in green's creation
By using StaticArrays.jl, using more `sincos(x)` and being more
conscious of the type stability of some functions, we add a roughly 33%
speedup in the creation of the green's operator. The creation of the
green's is also more efficiently parallelized, so more threads should
improve the creation of the green's operator.